Undo breakage of build.
git-svn-id: http://svn.code.sf.net/p/gratia/svn/trunk@4230 c85a5f45-a5f3-44fc-b02c-7891c381c519

+++ collector/gratia-services/net/sf/gratia/services/RecordProcessor.java
@@ -51,7 +51,7 @@ public class RecordProcessor extends Thread {
    ErrorRecorder errorRecorder = new ErrorRecorder();
    final Object lock;
    String historypath = "";
-   Hashtable<String,Integer> fProbeDetails = new Hashtable<String,Integer>();
+   Hashtable<String,Long> fProbeDetails = new Hashtable<String,Long>();
    static Pattern duplicateExceptionFinder = Pattern.compile("\\b[Dd]uplicate\\b");
    static Pattern metaFinder = Pattern.compile("_Meta ");
    static Pattern originFinder = Pattern.compile("Origin ");
